03_Nav/requirements.txt
numpy>=1.24
websockets>=12.0
pyserial>=3.5
